nest the message handler in the utils\discordHelpers
