+++ index.js
@@ -51,7 +51,6 @@ module.exports.appConfig = appConfig;
 const Discord = require('discord.js');
 const Consumer = require('sqs-consumer');
 const AWS = require('aws-sdk');
-const roles = require('./roles');
 const utils = require('./utils/discordHelpers');
 
 require('log-timestamp');
@@ -125,103 +124,11 @@ bot.on('ready', () => {
   }
 });
 
-function messageHandler(message) {
-  // Ignore bot messages
-  if (message.author.bot) {
-    return;
-  }
-
-  // Commands start with '!'
-  if (message.content[0] !== '!') {
-    return;
-  }
-
-  const commandText = message.content.split(' ')[0].substring(1).toLowerCase();
-  const command = commands[commandText];
-
-  // Check that the command exists
-  if (!command) {
-    return;
-  }
-
-  // If a command isn't allowed in a DM (or doesn't have allowDM defined),
-  // make sure we're in a guild.
-  if (!command.allowDM && !message.guild) {
-    message.reply('Sorry, I can only do that on a server. :frowning2:');
-    return;
-  }
-
-  // Checks that are only needed on a server
-  if (message.guild) {
-    // Check that the user is allowed to use the bot
-    let shouldIgnoreMessage = true;
-
-    // Check that the bot has any required roles at all
-    if (roles.REQUIRED_TO_USE_BOT.length > 0) {
-      // Try to find a common role between the required list and the
-      // user's roles
-      roles.REQUIRED_TO_USE_BOT.forEach((requiredRole) => {
-        if (message.member.roles.findKey('name', requiredRole)) {
-          shouldIgnoreMessage = false;
-        }
-      });
-    } else {
-      shouldIgnoreMessage = false;
-    }
-
-    // Check that the user is not part of a role that is banned from bot usage
-    roles.BANNED_FROM_BOT.forEach((bannedRole) => {
-      if (message.member.roles.findKey('name', bannedRole)) {
-        shouldIgnoreMessage = true;
-      }
-    });
-
-    if (shouldIgnoreMessage) {
-      return;
-    }
-
-    // If the command can only be used in certain channels, check that we're in
-    // one of those channels
-    if (command.onlyIn && command.onlyIn.length > 0) {
-      if (!utils.commandValidInChannel(command, message)) {
-        return;
-      }
-    }
-  }
-
-  // If the command requires roles, check that the user has one of them
-  if (command.requireRoles) {
-    // A command can't require roles and support DMs.
-    // This is a programmer error.
-    if (!message.guild) {
-      // TODO: Programmer error
-      return;
-    }
-
-    let satisfiesRoles = false;
-
-    // Loop through the roles needed by the command and see if the user
-    // has any of them.
-    command.requireRoles.forEach((role) => {
-      if (message.member.roles.findKey('name', role)) {
-        satisfiesRoles = true;
-      }
-    });
-
-    if (!satisfiesRoles) {
-      message.channel.send('I\'m sorry ' + message.author + ', I\'m ' +
-        'afraid I can\'t do that.');
-      return;
-    }
-  }
-
-  command.process(bot, message);
-}
 
 // Handle messages
 bot.on('message', message => {
   try {
-    messageHandler(message);
+    utils.messageHandler(bot, message);
   } catch (e) {
     console.error(e.stack);
   }
